Split admin helpers and migration-status out of functions-admin
functions-admin.ts had grown past the ~250 LOC target. Extract the
Firestore IO helpers (deleteUserExerciseData, readUserActivity) into
admin/user-data-ops.ts and move the migration-status callables into
functions-migration-status.ts, bringing functions-admin.ts to ~207 LOC.
Also corrects readUserActivity's doc comment, which still described the
old chunked `in`-query read instead of the getAll fetch.
